+++ SmartDeviceLink/SDLSetMediaClockTimer.h
@@ -31,7 +31,8 @@
  */
 - (instancetype)initWithDictionary:(NSMutableDictionary *)dict;
 
-- (instancetype)initWithUpdateMode:(SDLUpdateMode *)updateMode hours:(UInt8)hours minutes:(UInt8)minutes seconds:(UInt8)seconds;
+// TODO: (Alex M.)[2016-12-1] Change NSInteger to UInt8
+- (instancetype)initWithUpdateMode:(SDLUpdateMode *)updateMode hours:(NSInteger)hours minutes:(NSInteger)minutes seconds:(NSInteger)seconds;
 
 - (instancetype)initWithUpdateMode:(SDLUpdateMode *)updateMode;
