+++ SmartDeviceLink/SDLSetMediaClockTimer.m
@@ -24,7 +24,7 @@
 }
 
 
-- (instancetype)initWithUpdateMode:(SDLUpdateMode *)updateMode hours:(UInt8)hours minutes:(UInt8)minutes seconds:(UInt8)seconds {
+- (instancetype)initWithUpdateMode:(SDLUpdateMode *)updateMode hours:(NSInteger)hours minutes:(NSInteger)minutes seconds:(NSInteger)seconds {
     self = [self initWithUpdateMode:updateMode];
     if (!self) {
         return nil;
